requirements.txt
discord.py
requests
psutil
aiohttp
pywin32
pyperclip 


